+++ tests/testthat/test-chk-in.R
@@ -33,3 +33,12 @@ test_that("chk_in", {
   expect_error(chk_in(3L, values = c(1L, 1L, 7L), x_name = "1"), 
                "^1 must match 1 or 7, not 3[.]$")
 })
+
+test_that("chk_has", {
+  expect_true(chk_has(1, 1))
+  expect_true(chk_has(1:2, 1))
+  expect_false(chk_has(1, 1:2, err = FALSE))
+  expect_error(chk_has(1, 1:2), "^`1` must have values matching 1 and 2[.]$")
+  expect_true(chk_has(1:3, 1:2))
+  expect_error(chk_has(2, 1), "^`2` must have a value matching 1[.]$")
+})