+++ NEWS.md
@@ -1,3 +1,4 @@
+- Added `chk_has()` function.
 - Added `n = NULL` and `call. = FALSE` args to error message generating funcions.
 - Added `ellipsis = 10L` argument to `cc()`.
 - Added `sep = ", "` arg to `cc()`.